+++ src/main.rs
@@ -1,7 +1,4 @@
 mod app;
-mod db;
-mod export;
-mod model;
 mod ui;
 
 use std::{io, panic};